Core: - Markers are not much faster - UnityDrawMeshProvider API changes: GetRootData() and SetRootData() methods added - Editor: Scripts creation templates added

+++ Assets/ECS/Example/Scripts/Entities/Unit.cs
@@ -14,7 +14,7 @@ namespace ME.Example.Game.Entities {
         public float speed;
         public Entity pointFrom;
         public Entity pointTo;
-        public int lifes;
+        public float lifes;
 
     }
